1) Fix bug with properties being passed around job flows -- Azkaban tried to resolve the properties too early, didn't allow for a job to return a value. 2) Add job-runner servlet -- Allows for an external entity to POST up a JSON object of override properties and have that fire off a specific job with those properties (sample url to POST to is /job-runner/job-to-run) 3) Add null check to CachingFlowManager to handle Flows that don't exist better than an NPE 4) Adjust javaprocess to allow for main arguments that have commas in them (main.args is now interpretted as the string of main arguments) 5) Add ability to supress success email messages from a job (new property "azkaban.send.success.email") 6) Initial steps for supporting tar files.  Unfortunately, Java doesn't have great tar support so I gave up halfway.  Checking in the code anyway because it makes it a bit easier to handle multiple file types. 7) A lot of changes to mr-kluj contrib project.

+++ contrib/mr-kluj/src/main/java/com/linkedin/mr_kluj/StagedOutputJob.java
@@ -17,6 +17,7 @@
 package com.linkedin.mr_kluj;
 
 import org.apache.hadoop.conf.Configuration;
+import org.apache.hadoop.fs.FileStatus;
 import org.apache.hadoop.fs.FileSystem;
 import org.apache.hadoop.fs.Path;
 import org.apache.hadoop.mapreduce.Job;
@@ -25,6 +26,7 @@ import org.apache.log4j.Logger;
 
 import java.io.IOException;
 import java.net.URLClassLoader;
+import java.util.Map;
 
 /**
  *
@@ -90,11 +92,24 @@ public class StagedOutputJob extends Job
 
             fs.mkdirs(actualOutputPath);
 
-            log.info(String.format("Deleting data at old path[%s]", actualOutputPath));
-            fs.delete(actualOutputPath, true);
+            if (getConfiguration().getBoolean("com.linkedin.mr_kluj.delete.output.path", true)) {
+                log.info(String.format("Deleting data at old path[%s]", actualOutputPath));
+                fs.delete(actualOutputPath, true);
+            }
 
-            log.info(String.format("Moving from staged path[%s] to final resting place[%s]", stagedPath, actualOutputPath));
-            return fs.rename(stagedPath, actualOutputPath);
+            for (FileStatus fileStatus : FSUtils.spiderPath(fs, stagedPath)) {
+                Path thisStagedPath = fileStatus.getPath();
+                Path thisActualOutputPath = new Path(fileStatus.getPath().toString().replace(stagedPath.toString(), actualOutputPath.toString()));
+
+                log.info(String.format("Moving from staged path[%s] to final resting place[%s]", thisStagedPath, thisActualOutputPath));
+                fs.mkdirs(thisActualOutputPath.getParent());
+                if (! fs.rename(thisStagedPath, thisActualOutputPath)) {
+                    log.info("Rename failed!");
+                    return false;
+                }
+            }
+
+            return true;
         }
 
         log.warn("retVal was false for some reason...");
